Move norm func
