+++ app/personal_values/routes.py
@@ -10,14 +10,7 @@ from app.models import Scores
 
 from app import auto
 
-
-def normalize_scores(values):
-    min_score = min([value["score"] for value in values])
-    max_score = max([value["score"] for value in values])
-    # TODO: Check this correct way to normalise scores
-    for val in values:
-        val["score"] = (val["score"] - min_score) / (max_score - min_score) * 10
-    return values
+from app.personal_values.normalize import normalize_scores
 
 
 @bp.route("/personal_values", methods=["GET"])
